db.1.85: fix error return from __bt_first
The function returns an int, so NULL does not make sense.
The NULL value corresponds to that of RET_SUCCESS,
but all the other RET_SUCCESS returns are preceded
by an assignment to *erval and the caller also depends
on this assignment.

Presumably, the missing record should be treated as an error,
just like the failing mpool_get further down.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>