db.1.85: fix return from hcreate
This function should return nonzero on success.
If the size of an int is smaller than that of a pointer,
the truncated pointer value could happen to be zero
despite the successful allocation.

Note that nvi does not call this function, but the code
is still getting compiled and could generate a warning.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>