add wcwidth prototype if not readily available
Technically, wcwidth is only declared by wchar.h
if _XOPEN_SOURCE is defined.
However, defining that hides a bunch of other symbols.
On some systems, _GNU_SOURCE would work too,
but that brings in too many symbols.

It is easier to just declare wcwidth explicitly when it is not declared
by default.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>