ex: properly handle (global) insert with trailing delimiter or newline
The comment above the modified if-statement suggests
that a trailing delimiter should cause an empty line
to be inserted and this is also what the version of vi
that comes with AIX and that identifies itself as "Version 3.10" does.

However, at this point there is no obvious way of determining
whether there was such a trailing delimiter since
the delimiter is removed in the caller (ex_cmd).
Keep track of this removal so that an empty line can be inserted
when this removal results in an empty trailing text.
Do the same for a trailing newline.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>