ex: properly global insert with initial newline in inserted text
The line insertion was guarded by a test that would fail
to trigger on empty lines in the inserted text.
It is not clear why this guard appears in the code.
Removing the guard reportedly makes the code behave
in the same way as the original vi.
The version of vi that comes with AIX and that identifies itself
as "Version 3.10" also behaves in this way.

Reported-by: Mohamed Akram <mohd.akram@outlook.com>
Patch-by: Mohamed Akram <mohd.akram@outlook.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>